sof-event-extractor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
spacy==3.7.2
openai==1.3.7
google-generativeai==0.3.2
scikit-learn==1.3.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
numpy==1.25.2
regex==2023.10.3
streamlit==1.37.1
pandas==2.2.2
plotly==5.23.0
numpy==1.26.4
pdfplumber==0.11.4
pymupdf==1.24.9
pillow==10.4.0
pytesseract==0.3.13
python-docx==1.1.2
dateparser==1.2.0
groq==0.9.0
httpx==0.27.2
python-dotenv==1.0.1
